Schema of Interface
