+++ app/app.js
@@ -1,4 +1,5 @@
 'use strict';
-const logic = require('./inquirer');
+
+const logic = require('./interface/logic');
 
 logic();
