+++ app/interface/main_menu.js
@@ -1,0 +1,23 @@
+
+const inquirer = require('inquirer');
+
+
+function renderMainMenu () {
+// TODO
+}
+
+function renderCreateTask () {
+// TODO
+}
+
+function renderListTasks () {
+// TODO
+}
+
+function renderSeeTask () {
+// TODO
+}
+
+
+
+module.exports = {renderMainMenu, renderCreateTask, renderListTasks, renderSeeTask};
